+++ src/main/kotlin/com/youquiz/quiz/dto/QuizResponse.kt
@@ -11,7 +11,7 @@ data class QuizResponse(
     val solution: String,
     val writer: User,
     val chapterId: String,
-    val answerRate: Long,
+    val answerRate: Double,
     val correctCount: Long,
     val incorrectCount: Long,
     val createdDate: LocalDateTime,
